見やすさのために準備関数を全てクラスごとに分ける
