+++ srcs/HttpRequest/HttpRequest.cpp
@@ -86,128 +86,6 @@ bool HttpRequest::is_requestlineformat(std::string input_requestline)
 	return (true);
 }
 
-LinkClass* HttpRequest::ready_LinkClass(std::map<std::string, std::map<std::string, std::string> > link_valuemap)
-{
-	return (new LinkClass(link_valuemap));
-}
-
-SecurityPolicy* HttpRequest::ready_SecurityPolicy(const std::string &report_url, std::map<std::string, std::vector<std::string> >	_policy_directive)
-{
-	return (new SecurityPolicy(report_url, _policy_directive));
-}
-
-SecurityPolicy* HttpRequest::ready_SecurityPolicy(std::map<std::string, std::vector<std::string> >	_policy_directive)
-{
-	return (new SecurityPolicy(_policy_directive));
-}
-
-TwoValueSet* HttpRequest::ready_TwoValueSet(const std::string &all_value)
-{
-	std::stringstream	ss(HandlingString::obtain_withoutows_value(all_value));
-	std::string			first_value;
-	std::string			second_value;
-
-	std::getline(ss, first_value, '/');
-	std::getline(ss, second_value, '/');
-
-	return (new TwoValueSet(first_value, second_value));
-}
-
-TwoValueSet* HttpRequest::ready_TwoValueSet(const std::string &value, char delimiter)
-{
-	std::stringstream	ss(HandlingString::obtain_withoutows_value(value));
-	std::string			first_value;
-	std::string			second_value;
-
-	std::getline(ss, first_value, delimiter);
-	std::getline(ss, second_value, delimiter);
-	return (new TwoValueSet(HandlingString::obtain_withoutows_value(first_value), HandlingString::obtain_withoutows_value(second_value)));
-}
-
-ValueArraySet* HttpRequest::ready_ValueArraySet(const std::string &all_value)
-{
-	std::vector<std::string>	value_array;
-	std::stringstream			ss(all_value);
-	std::string					line;
-
-	while(std::getline(ss, line, ','))
-		value_array.push_back(HandlingString::obtain_withoutows_value(line));
-	return (new ValueArraySet(value_array));
-}
-
-ValueDateSet* HttpRequest::ready_ValueDateSet(const std::string &value)
-{
-	return (new ValueDateSet(HandlingString::obtain_withoutows_value(value)));
-}
-
-ValueMap* HttpRequest::ready_ValueMap(const std::string &value, char delimiter)
-{
-	std::map<std::string, std::string>	value_map;
-	std::stringstream					ss(value);
-	std::string							line;
-
-	while(std::getline(ss, line, delimiter))
-		value_map[HandlingString::obtain_word_before_delimiter(HandlingString::obtain_withoutows_value(line), '=')] \
-		= HandlingString::obtain_word_after_delimiter(HandlingString::obtain_withoutows_value(line), '=');
-	return (new ValueMap(value_map));
-}
-
-ValueMap* HttpRequest::ready_ValueMap(const std::string &value)
-{
-	std::map<std::string, std::string> value_map;
-	std::stringstream	ss(value);
-	std::string			line;
-
-	while(std::getline(ss, line, ';'))
-		value_map[HandlingString::obtain_word_before_delimiter(HandlingString::obtain_withoutows_value(line), '=')] \
-		= HandlingString::obtain_word_after_delimiter(HandlingString::obtain_withoutows_value(line), '=');
-	return (new ValueMap(value_map));
-}
-
-ValueMap* HttpRequest::ready_ValueMap(const std::string &only_value, const std::string &value)
-{
-	std::map<std::string, std::string>	value_map;
-	std::stringstream					ss(value);
-	std::string							line;
-	std::string							skipping_word;
-
-	while(std::getline(ss, line, ';'))
-	{
-		skipping_word = HandlingString::obtain_withoutows_value(line);
-		value_map[HandlingString::obtain_word_before_delimiter(skipping_word, '=')] \
-		= HandlingString::obtain_withoutows_value(HandlingString::obtain_word_after_delimiter(skipping_word, '='));
-	}
-	return (new ValueMap(only_value, value_map));
-}
-
-ValueSet* HttpRequest::ready_ValueSet(const std::string &value)
-{
-	return (new ValueSet(HandlingString::obtain_withoutows_value(value)));
-}
-
-ValueWeightArraySet*	HttpRequest::ready_ValueWeightArraySet(const std::string &value)
-{
-	std::map<std::string, double>	value_map;
-	std::stringstream				splited_by_commma(value);
-	std::string						line;
-	std::string						changed_line;
-	std::string						target_value;
-
-	while(std::getline(splited_by_commma, line, ','))
-	{
-		changed_line = HandlingString::obtain_withoutows_value(line);
-		if (changed_line.find(';') != std::string::npos)
-		{
-			target_value = HandlingString::obtain_weight(HandlingString::obtain_word_after_delimiter(changed_line, ';'));
-			value_map[HandlingString::obtain_word_before_delimiter(changed_line, ';')] = \
-			HandlingString::str_to_double(HandlingString::obtain_weight(HandlingString::obtain_word_after_delimiter(changed_line, ';')));
-		}
-		else
-			value_map[changed_line] = 1.0;
-	}
-	return (new ValueWeightArraySet(value_map));
-}
-
 bool	HttpRequest::is_keyword_exist(const std::string &key)
 {
 	if (this->_field_name_parser.count(key) > 0)
@@ -235,7 +113,16 @@ bool	HttpRequest::is_requestformat(const std::string &val)
 	return (true);
 }
 
-// Accept-Encoding: gzip;q=0.5
+std::string	HttpRequest::obtain_request_value(const std::string value)
+{
+	std::string::size_type pos = value.find_first_of(":");
+	std::string part2 = value.substr(pos + 2);
+
+	return (part2);
+}
+
+// weightarrayset わかりやすいように
+
 bool	HttpRequest::is_weightformat(const std::string &value)
 {
 	size_t		semicolon_pos = value.find(';');
@@ -252,330 +139,11 @@ bool	HttpRequest::is_weightformat(const std::string &value)
 	return HandlingString::is_double(weight_num);
 }
 
-std::string	HttpRequest::obtain_request_value(const std::string value)
-{
-	std::string::size_type pos = value.find_first_of(":");
-	std::string part2 = value.substr(pos + 2);
+// linkclass
 
-	return (part2);
-}
+//~~~~
 
-void HttpRequest::set_accept(const std::string &key, const std::string &value)
-{
-	size_t							value_length = value.size();
-	size_t							now_location = 0;
-	std::stringstream				splited_by_commma(value);
-	std::string						line;
-	std::string						changed_line;
-
-	char not_accept_encoding_keyset[] = {
-		'\x00', '\x01', '\x02', '\x03', '\x04', '\x05', '\x06', '\x07', '\x08', '\x09', '\x10', \
-		'\x0A', '\x0B', '\x0C', '\x0D', '\x0E', '\x0F', '\x10', '\x11', '\x12', '\x13', '\x14', \
-		'\x15', '\x16', '\x17', '\x18', '\x19', '\x1A', '\x1B', '\x1C', '\x1D', '\x1E', '\x1F', \
-		'(', ')', ':', '<', '>', '?', '@', '[', '\\', ']', '{', '}'
-	};
-	const std::set<char> not_accept_keyset
-	(
-		not_accept_encoding_keyset,
-		not_accept_encoding_keyset + sizeof(not_accept_encoding_keyset) / sizeof(not_accept_encoding_keyset[0])
-	);
-	while (now_location != value_length - 1)
-	{
-		if (not_accept_keyset.count(value[now_location]) > 0)
-			return;
-		now_location++;
-	}
-	while(std::getline(splited_by_commma, line, ','))
-	{
-		if (line[0] == ' ')
-			changed_line = line.substr(1);
-		else
-			changed_line = line;
-		if (changed_line.find(';') != std::string::npos)
-		{
-			if (HandlingString::is_double(HandlingString::obtain_weight(HandlingString::obtain_word_after_delimiter(changed_line, ';'))) == false)
-				return;
-		}
-	}
-	this->_request_keyvalue_map[key] = ready_ValueWeightArraySet(value);
-}
-
-void	HttpRequest::set_accept_ch(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueArraySet(value);
-}
-
-void	HttpRequest::set_accept_charset(const std::string &key, const std::string &value)
-{
-	std::stringstream				splited_by_commma(value);
-	std::string						line;
-	std::string						changed_line;
-
-	while(std::getline(splited_by_commma, line, ','))
-	{
-		if (line[0] == ' ')
-			changed_line = line.substr(1);
-		else
-			changed_line = line;
-		if (changed_line.find(';') != std::string::npos)
-		{
-			if (HandlingString::is_double(HandlingString::obtain_weight(HandlingString::obtain_word_after_delimiter(changed_line, ';'))) == false)
-				return;
-		}
-	}
-	this->_request_keyvalue_map[key] = ready_ValueWeightArraySet(value);
-}
-
-void	HttpRequest::set_accept_encoding(const std::string &key, const std::string &value)
-{
-	std::stringstream 	splited_by_commma(value);
-	std::string			skipping_nokeyword;
-	std::string			keyword;
-	std::string			line;
-	const std::string accept_encoding_keyset[] = {
-		"gzip", "compress", "deflate", "br", "*", "identity"
-	};
-	const std::set<std::string> httprequest_keyset
-	(
-		accept_encoding_keyset,
-		accept_encoding_keyset + sizeof(accept_encoding_keyset) / sizeof(accept_encoding_keyset[0])
-	);
-
-	while(std::getline(splited_by_commma, line, ','))
-	{
-		if (HandlingString::obtain_withoutows_value(line) != "")
-		{
-			if (line.find(';') != std::string::npos)
-			{
-				if (std::count(line.begin(), line.end(), ';') != 1 || this->is_weightformat(line) == false)
-				{
-					this->_status_code = 400;
-					return;
-				}
-				keyword = HandlingString::obtain_word_before_delimiter(line, ';');
-			}
-			else
-				keyword = line;
-			keyword = HandlingString::obtain_withoutows_value(keyword);
-			if (httprequest_keyset.count(keyword) > 0)
-				skipping_nokeyword = skipping_nokeyword + line + ',';
-		}
-	}
-	this->_request_keyvalue_map[key] = ready_ValueWeightArraySet(skipping_nokeyword.substr(0, skipping_nokeyword.length() - 1));
-}
-
-void	HttpRequest::set_accept_language(const std::string &key, const std::string &value)
-{
-	std::stringstream splited_by_commma(value);
-	std::string	skipping_nokeyword;
-	std::string	accept_language_value;
-	std::string	line;
-	std::string	target_value;
-
-	while(std::getline(splited_by_commma, line, ','))
-	{
-		if (line.find(';') != std::string::npos)
-		{
-			accept_language_value = HandlingString::obtain_word_before_delimiter(line, ';');
-			target_value = HandlingString::obtain_weight(HandlingString::obtain_word_after_delimiter(line, ';'));
-			if (is_accept_langage_valueword(accept_language_value) == true)
-				skipping_nokeyword = skipping_nokeyword + line;
-			if (target_value == "0" || target_value == "0.0")
-				return;
-			if (HandlingString::is_double(target_value) == false)
-				return;
-		}
-		else
-		{
-			if (is_accept_langage_valueword(line) == true)
-				skipping_nokeyword = skipping_nokeyword + line;
-		}
-	}
-	this->_request_keyvalue_map[key] = ready_ValueWeightArraySet(value);
-}
-
-bool	HttpRequest::is_accept_langage_valueword(const std::string &value)
-{
-	size_t	string_length = value.length();
-	size_t	now_location = 0;
-
-	while (string_length != now_location)
-	{
-		if (!('0' <= value[now_location] && value[now_location] <= '9'))
-		{
-			if (!('A' <= value[now_location] && value[now_location] <= 'Z'))
-			{
-				if (!('a' <= value[now_location] && value[now_location] <= 'z'))
-				{
-					if (!(' ' == value[now_location] || '*' == value[now_location] || '-' == value[now_location] \
-					|| ',' == value[now_location] || '.' == value[now_location] || ';' == value[now_location] \
-					|| '=' == value[now_location]))
-					{
-						this->_status_code = 406;
-						return (false);
-					}
-				}
-			}
-		}
-		now_location++;
-	}
-	return (true);
-}
-
-// Accept-Patchどういう持ち方かわからん
-
-void	HttpRequest::set_accept_post(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_TwoValueSet(value, ',');
-}
-
-void	HttpRequest::set_accept_ranges(const std::string &key, const std::string &value)
-{
-	if (value == "bytes")
-		this->_request_keyvalue_map[key] = ready_ValueSet(value);
-	else
-		return;
-}
-
-void	HttpRequest::set_access_control_allow_credentials(const std::string &key, const std::string &value)
-{
-	if (value != "true")
-		return;
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_access_control_allow_headers(const std::string &key, const std::string &value)
-{
-	std::vector<std::string>	value_array;
-	std::stringstream	ss(value);
-	std::string			line;
-
-	while(std::getline(ss, line, ','))
-	{
-		if (this->is_keyword_exist(HandlingString::obtain_withoutows_value(line)) == false)
-			return;
-	}
-	this->_request_keyvalue_map[key] = ready_ValueArraySet(value);
-}
-
-void	HttpRequest::set_access_control_allow_methods(const std::string &key, const std::string &value)
-{
-	std::stringstream	ss(value);
-	std::string			line;
-	std::string			word;
-
-	while(std::getline(ss, line, ','))
-	{
-		if (line[0] == ' ')
-			word = line.substr(1);
-		else
-			word = line;
-		if (word != "GET" && word != "HEAD" && word != "POST" && word != "PUT" && word != "PUT" && word != "DELETE" \
-		&& word != "CONNECT" && word != "OPTIONS" && word != "TRACE" && word != "PATCH")
-			return;
-	}
-	this->_request_keyvalue_map[key] = ready_ValueArraySet(value);
-}
-
-void	HttpRequest::set_access_control_allow_origin(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_access_control_expose_headers(const std::string &key, const std::string &value)
-{
-	std::vector<std::string>	value_array;
-	std::stringstream	ss(value);
-	std::string			line;
-
-	while(std::getline(ss, line, ','))
-	{
-		if (this->is_keyword_exist(HandlingString::obtain_withoutows_value(line)) == false)
-			return;
-	}
-	this->_request_keyvalue_map[key] = ready_ValueArraySet(value);
-}
-
-void	HttpRequest::set_access_control_max_age(const std::string &key, const std::string &value)
-{
-	if (HandlingString::is_positive_and_under_intmax(value) == false)
-		return;
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_access_control_request_headers(const std::string &key, const std::string &value)
-{
-	std::vector<std::string>	value_array;
-	std::stringstream	ss(value);
-	std::string			line;
-	std::string			word;
-
-	while(std::getline(ss, line, ','))
-	{
-		if (line[0] == ' ')
-			word = line.substr(1);
-		else
-			word = line;
-		if (this->is_keyword_exist(HandlingString::obtain_withoutows_value(word)) == false)
-			return;
-	}
-	this->_request_keyvalue_map[key] = ready_ValueArraySet(value);
-}
-
-void	HttpRequest::set_access_control_request_method(const std::string &key, const std::string &value)
-{
-	if (value != "GET" && value != "HEAD" && value != "POST" && value != "PUT" && value != "PUT" && value != "DELETE" \
-		&& value != "CONNECT" && value != "OPTIONS" && value!= "TRACE" && value != "PATCH")
-		return;
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_age(const std::string &key, const std::string &value)
-{
-	if (HandlingString::is_positive_and_under_intmax(value) == false)
-		return;
-	if (HandlingString::is_positive_and_under_intmax(value) == false)
-		return;
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_allow(const std::string &key, const std::string &value)
-{
-	std::stringstream	ss(value);
-	std::string			line;
-	std::string			word;
-
-	while(std::getline(ss, line, ','))
-	{
-		if (line[0] == ' ')
-			word = line.substr(1);
-		else
-			word = line;
-		if (word != "GET" && word != "HEAD" && word != "POST" && word != "PUT" && word != "PUT" && word != "DELETE" \
-		&& word != "CONNECT" && word != "OPTIONS" && word != "TRACE" && word != "PATCH")
-			return;
-	}
-	this->_request_keyvalue_map[key] = ready_ValueArraySet(value);
-}
-
-void	HttpRequest::set_alt_svc(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueMap(value);
-}
-
-void	HttpRequest::set_alt_used(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_authorization(const std::string &key, const std::string &value)
-{
-	// Digest username=<username>,realm="<realm>",uri="<url>",algorithm=<algorithm>,nonce="<nonce>",
-	// ValueMapに変更
-	this->_request_keyvalue_map[key] = ready_TwoValueSet(value, ' ');
-}
-
-// Cache-Controlどう使うのか全くわからない
+// まだできてない
 void	HttpRequest::set_cache_control(const std::string &key, const std::string &value)
 {
 	(void)key;
@@ -583,113 +151,6 @@ void	HttpRequest::set_cache_control(const std::string &key, const std::string &v
 	// Digest username=<username>,realm="<realm>",uri="<url>",algorithm=<algorithm>,nonce="<nonce>",
 	// ValueMapに変更
 	// this->_request_keyvalue_map[key] = ready_ValueWeightArraySet(value);
-}
-
-void	HttpRequest::set_clear_site_data(const std::string &key, const std::string &value)
-{
-	// ダブルクオーテーションで囲う必要性があるようだが、"aaaa"", "bbb"みたいなことをされたとする
-	// この場合にフォーマットが正しくないみたいなステータスコード を投げる？
-	this->_request_keyvalue_map[key] = ready_ValueArraySet(value);
-}
-
-void	HttpRequest::set_connection(const std::string &key, const std::string &value)
-{
-	if (value == "close" || value == "keep-alive")
-		this->_request_keyvalue_map[key] = ready_ValueSet(value);
-	else
-		return;
-}
-
-void	HttpRequest::set_content_disponesition(const std::string &key, const std::string &value)
-{
-	std::stringstream	ss(value);
-	std::string			only_value;
-	std::string			except_onlyvalue_line;
-	std::string 		line;
-
-	std::getline(ss, only_value, ';');
-	while (std::getline(ss, line, ';'))
-		except_onlyvalue_line = except_onlyvalue_line + line;
-	this->_request_keyvalue_map[key] = ready_ValueMap(only_value, except_onlyvalue_line);
-}
-
-void	HttpRequest::set_content_encoding(const std::string &key, const std::string &value)
-{
-	std::stringstream	ss(value);
-	std::string			line;
-
-	while(std::getline(ss, line, ','))
-	{
-		if (line != "gzip" && line != "compress" && line != "deflate" && line != "br")
-			return;
-	}
-	this->_request_keyvalue_map[key] = ready_ValueArraySet(value);
-}
-
-void	HttpRequest::set_content_language(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueArraySet(value);
-}
-
-void	HttpRequest::set_content_length(const std::string &key, const std::string &value)
-{
-	if (HandlingString::is_positive_and_under_intmax(value) == false)
-		return;
-	if (HandlingString::is_positive_and_under_intmax(value) == false)
-		return;
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_content_location(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_content_range(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-std::vector<std::string>	HttpRequest::securitypolicy_readyvector(const std::string &words)
-{
-	std::vector<std::string>	tmp_vector;
-	std::string	skippingfirstemptyword;
-	std::string			line;
-
-	if (words[0] == ' ')
-		skippingfirstemptyword = words.substr(1);
-	else
-		skippingfirstemptyword = words;
-
-	std::stringstream	ss(skippingfirstemptyword);
-
-	while(std::getline(ss, line, ' '))
-		tmp_vector.push_back(line);
-	return (tmp_vector);
-}
-
-void	HttpRequest::set_content_security_policy(const std::string &key, const std::string &value)
-{
-	std::map<std::string, std::vector<std::string> > content_security_map;
-	std::stringstream	ss(value);
-	std::string			line;
-	std::string			skip_emptyword;
-	std::string			policy_directive;
-	std::string			words;
-	size_t				empty_position;
-
-	while(std::getline(ss, line, ';'))
-	{
-		if (line[0] == ' ')
-			skip_emptyword = line.substr(1);
-		else
-			skip_emptyword = line;
-		empty_position = skip_emptyword.find(' ');
-		policy_directive = skip_emptyword.substr(0, empty_position);
-		words = skip_emptyword.substr(empty_position + 1);
-		content_security_map[policy_directive] = securitypolicy_readyvector(words);
-	}
-	this->_request_keyvalue_map[key] = ready_SecurityPolicy(content_security_map);
 }
 
 void	HttpRequest::set_content_security_policy_report_only(const std::string &key, const std::string &value)
@@ -700,255 +161,18 @@ void	HttpRequest::set_content_security_policy_report_only(const std::string &key
 	// this->_request_keyvalue_map[key] = ready_SecurityPolicy();
 }
 
-void	HttpRequest::set_content_type(const std::string &key, const std::string &value)
+void	HttpRequest::set_servertiming(const std::string &key, const std::string &value)
 {
-	std::stringstream	ss(value);
-	std::string			only_value;
-	std::string			except_onlyvalue_line;
-	std::string 		line;
-
-	std::getline(ss, only_value, ',');
-	while (std::getline(ss, line, ','))
-		except_onlyvalue_line = except_onlyvalue_line + line;
-	this->_request_keyvalue_map[key] = ready_ValueMap(only_value, except_onlyvalue_line);
+	// cpu;dur=2.4;a=b, cpu; ,,,みたいな感じなのでmapで保持しないほうがいいかもしれない
+	// this->_request_keyvalue_map[key] = ready_ValueMap(value);
+	(void)key;
+	(void)value;
 }
 
-void	HttpRequest::set_cookie(const std::string &key, const std::string &value)
+void	HttpRequest::set_x_xss_protection(const std::string &key, const std::string &value)
 {
-	this->_request_keyvalue_map[key] = ready_ValueMap(value);
-}
-
-void	HttpRequest::set_cross_origin_embedder_policy(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_cross_origin_opener_policy(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_cross_origin_resource_policy(const std::string &key, const std::string &value)
-{
-	if (value == "same-site" || value == "same-origin" || value == "cross-origin")
-		this->_request_keyvalue_map[key] = ready_ValueSet(value);
-	else
-		return;
-}
-
-void	HttpRequest::set_date(const std::string &key, const std::string &value)
-{
-	std::stringstream	ss(value);
-	std::string			line;
-	std::string			day_name;
-	std::string			day;
-	std::string			month;
-	std::string			year;
-	std::string			hour;
-	std::string			minute;
-	std::string			second;
-
-	std::getline(ss, day_name, ',');
-	std::getline(ss, line, ',');
-	if (day_name != "Mon" && day_name != "Tue" && day_name != "Wed" && day_name != "Thu" && day_name != "Fri" && day_name != "Sat" && day_name != "Sun")
-		return;
-	std::string after_line = line.substr(1);
-	std::stringstream	sss(after_line);
-	std::getline(sss, day, ' ');
-	if (day.length() != 2)
-		return;
-	if (!(1 <= HandlingString::str_to_int(day) || HandlingString::str_to_int(day) <= 31))
-		return;
-	std::getline(sss, month, ' ');
-	std::getline(sss, year, ' ');
-	std::string			hour_minute_second;
-	std::getline(sss, hour_minute_second, ' ');
-	std::stringstream	ssss(hour_minute_second);
-	std::getline(ssss, hour, ':');
-	if (!(0 <= HandlingString::str_to_int(hour) || HandlingString::str_to_int(hour) <= 60))
-		return;
-	std::getline(ssss, minute, ':');
-	if (!(0 <= HandlingString::str_to_int(minute) || HandlingString::str_to_int(minute) <= 60))
-		return;
-	std::getline(ssss, second, ':');
-	if (!(0 <= HandlingString::str_to_int(second) || HandlingString::str_to_int(second) <= 60))
-		return;
-	this->_request_keyvalue_map[key] = ready_ValueDateSet(value);
-}
-
-void	HttpRequest::set_etag(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_expect(const std::string &key, const std::string &value)
-{
-	if (value == "100-continue")
-		this->_request_keyvalue_map[key] = ready_ValueSet(value);
-	else
-		return;
-}
-
-// expect-ctは現状廃止されているっぽくて対応したくない
-
-void	HttpRequest::set_expires(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_forwarded(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueMap(value);
-}
-
-void	HttpRequest::set_from(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_host(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_TwoValueSet(value);
-}
-
-void	HttpRequest::set_if_match(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueArraySet(value);
-}
-
-void	HttpRequest::set_if_modified_since(const std::string &key, const std::string &value)
-{
-	std::stringstream	ss(value);
-	std::string			line;
-	std::string			day_name;
-	std::string			day;
-	std::string			month;
-	std::string			year;
-	std::string			hour;
-	std::string			minute;
-	std::string			second;
-
-	std::getline(ss, day_name, ',');
-	std::getline(ss, line, ',');
-	if (day_name != "Mon" && day_name != "Tue" && day_name != "Wed" && day_name != "Thu" && day_name != "Fri" && day_name != "Sat" && day_name != "Sun")
-		return;
-	std::string after_line = line.substr(1);
-	std::stringstream	sss(after_line);
-	std::getline(sss, day, ' ');
-	if (day.length() != 2)
-		return;
-	if (!(1 <= HandlingString::str_to_int(day) || HandlingString::str_to_int(day) <= 31))
-		return;
-	std::getline(sss, month, ' ');
-	std::getline(sss, year, ' ');
-	std::string			hour_minute_second;
-	std::getline(sss, hour_minute_second, ' ');
-	std::stringstream	ssss(hour_minute_second);
-	std::getline(ssss, hour, ':');
-	if (!(0 <= HandlingString::str_to_int(hour) || HandlingString::str_to_int(hour) <= 60))
-		return;
-	std::getline(ssss, minute, ':');
-	if (!(0 <= HandlingString::str_to_int(minute) || HandlingString::str_to_int(minute) <= 60))
-		return;
-	std::getline(ssss, second, ':');
-	if (!(0 <= HandlingString::str_to_int(second) || HandlingString::str_to_int(second) <= 60))
-		return;
-	this->_request_keyvalue_map[key] = ready_ValueDateSet(value);
-}
-
-void	HttpRequest::set_if_none_match(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueArraySet(value);
-}
-
-void	HttpRequest::set_if_range(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_if_unmodified_since(const std::string &key, const std::string &value)
-{
-	std::stringstream	ss(value);
-	std::string			line;
-	std::string			day_name;
-	std::string			day;
-	std::string			month;
-	std::string			year;
-	std::string			hour;
-	std::string			minute;
-	std::string			second;
-
-	std::getline(ss, day_name, ',');
-	std::getline(ss, line, ',');
-	if (day_name != "Mon" && day_name != "Tue" && day_name != "Wed" && day_name != "Thu" && day_name != "Fri" && day_name != "Sat" && day_name != "Sun")
-		return;
-	std::string after_line = line.substr(1);
-	std::stringstream	sss(after_line);
-	std::getline(sss, day, ' ');
-	if (day.length() != 2)
-		return;
-	if (!(1 <= HandlingString::str_to_int(day) || HandlingString::str_to_int(day) <= 31))
-		return;
-	std::getline(sss, month, ' ');
-	std::getline(sss, year, ' ');
-	std::string			hour_minute_second;
-	std::getline(sss, hour_minute_second, ' ');
-	std::stringstream	ssss(hour_minute_second);
-	std::getline(ssss, hour, ':');
-	if (!(0 <= HandlingString::str_to_int(hour) || HandlingString::str_to_int(hour) <= 60))
-		return;
-	std::getline(ssss, minute, ':');
-	if (!(0 <= HandlingString::str_to_int(minute) || HandlingString::str_to_int(minute) <= 60))
-		return;
-	std::getline(ssss, second, ':');
-	if (!(0 <= HandlingString::str_to_int(second) || HandlingString::str_to_int(second) <= 60))
-		return;
-	this->_request_keyvalue_map[key] = ready_ValueDateSet(value);
-}
-
-void	HttpRequest::set_keep_alive(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueMap(value, ',');
-}
-
-void	HttpRequest::set_last_modified(const std::string &key, const std::string &value)
-{
-	std::stringstream	ss(value);
-	std::string			line;
-	std::string			day_name;
-	std::string			day;
-	std::string			month;
-	std::string			year;
-	std::string			hour;
-	std::string			minute;
-	std::string			second;
-
-	std::getline(ss, day_name, ',');
-	std::getline(ss, line, ',');
-	if (day_name != "Mon" && day_name != "Tue" && day_name != "Wed" && day_name != "Thu" && day_name != "Fri" && day_name != "Sat" && day_name != "Sun")
-		return;
-	std::string after_line = line.substr(1);
-	std::stringstream	sss(after_line);
-	std::getline(sss, day, ' ');
-	if (day.length() != 2)
-		return;
-	if (!(1 <= HandlingString::str_to_int(day) || HandlingString::str_to_int(day) <= 31))
-		return;
-	std::getline(sss, month, ' ');
-	std::getline(sss, year, ' ');
-	std::string			hour_minute_second;
-	std::getline(sss, hour_minute_second, ' ');
-	std::stringstream	ssss(hour_minute_second);
-	std::getline(ssss, hour, ':');
-	if (!(0 <= HandlingString::str_to_int(hour) || HandlingString::str_to_int(hour) <= 60))
-		return;
-	std::getline(ssss, minute, ':');
-	if (!(0 <= HandlingString::str_to_int(minute) || HandlingString::str_to_int(minute) <= 60))
-		return;
-	std::getline(ssss, second, ':');
-	if (!(0 <= HandlingString::str_to_int(second) || HandlingString::str_to_int(second) <= 60))
-		return;
-	this->_request_keyvalue_map[key] = ready_ValueDateSet(value);
+	(void)key;
+	(void)value;
 }
 
 std::map<std::string, std::string>	HttpRequest::ready_mappingvalue(const std::string &value_map_words)
@@ -967,261 +191,6 @@ std::map<std::string, std::string>	HttpRequest::ready_mappingvalue(const std::st
 		words_mapping[key] = val;
 	}
 	return (words_mapping);
-}
-
-void	HttpRequest::set_link(const std::string &key, const std::string &value)
-{
-	std::map<std::string, std::map<std::string, std::string> > value_map;
-	std::stringstream	ss(value);
-	std::string			line;
-	std::string			uri;
-	std::string			mapping_value;
-
-	while(std::getline(ss, line, ','))
-	{
-		uri = HandlingString::obtain_word_before_delimiter(line, ';');
-		mapping_value = HandlingString::obtain_word_after_delimiter(line, ';');
-		value_map[uri] = this->ready_mappingvalue(mapping_value);
-	}
-	this->_request_keyvalue_map[key] = ready_LinkClass(value_map);
-}
-
-void	HttpRequest::set_location(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_max_forwards(const std::string &key, const std::string &value)
-{
-	if (HandlingString::is_positive_and_under_intmax(value))
-		this->_request_keyvalue_map[key] = ready_ValueSet(value);
-	else
-		return;
-}
-
-void	HttpRequest::set_origin(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_permission_policy(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_TwoValueSet(value, ',');
-}
-
-void	HttpRequest::set_proxy_authenticate(const std::string &key, const std::string &value)
-{
-	std::string re_line = value.substr(1);
-	size_t	empty_position = re_line.find(' ');
-	std::string	before_space_word = re_line.substr(0, empty_position);
-	std::string	after_space_word = re_line.substr(empty_position + 1, re_line.length());
-
-	this->_request_keyvalue_map[key] = ready_ValueMap(before_space_word, after_space_word);
-}
-
-void	HttpRequest::set_proxy_authorization(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_TwoValueSet(value, ' ');
-}
-
-// range何かよくわからん
-
-void	HttpRequest::set_referer(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_referrer_policy(const std::string &key, const std::string &value)
-{
-	if (value == "no-referrer" || value == "no-referrer-when-downgrade" || value == "origin" || value == "origin-when-cross-origin" || \
-	value == "same-origin" || value == "strict-origin" || value == "strict-origin-when-cross-origin" || value == "unsafe-url")
-		this->_request_keyvalue_map[key] = ready_ValueSet(value);
-	else
-		return;
-}
-
-void	HttpRequest::set_retry_after(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_sec_fetch_dest(const std::string &key, const std::string &value)
-{
-	if (value == "audio" || value == "audioworklet" || value == "document" || value == "embed" || \
-	value == "empty" || value == "font" || value == "frame" || value == "iframe" || value == "image" || value == "manifest" || \
-	value == "object" || value == "paintworklet" || value == "report" || value == "script" || value == "serviceworker" || \
-	value == "sharedworker" || value == "style" || value == "track" || value == "video" || value == "worker" || value == "xslt")
-		this->_request_keyvalue_map[key] = ready_ValueSet(value);
-	else
-		return;
-}
-
-void	HttpRequest::set_sec_fetch_mode(const std::string &key, const std::string &value)
-{
-	if (value == "cors" || value == "navigate" || value == "no-cors" || value == "same-origin" || value == "websocket")
-		this->_request_keyvalue_map[key] = ready_ValueSet(value);
-	else
-		return;
-}
-
-void	HttpRequest::set_sec_fetch_site(const std::string &key, const std::string &value)
-{
-	if (value == "cross-site" || value == "same-origin" || value == "same-site" || value == "none")
-		this->_request_keyvalue_map[key] = ready_ValueSet(value);
-	else
-		return;
-}
-
-void	HttpRequest::set_sec_fetch_user(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_sec_purpose(const std::string &key, const std::string &value)
-{
-	if (value == "prefetch")
-		this->_request_keyvalue_map[key] = ready_ValueSet(value);
-	else
-		return;
-}
-
-void	HttpRequest::set_sec_websocket_accept(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_server(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_servertiming(const std::string &key, const std::string &value)
-{
-	// cpu;dur=2.4;a=b, cpu; ,,,みたいな感じなのでmapで保持しないほうがいいかもしれない
-	// this->_request_keyvalue_map[key] = ready_ValueMap(value);
-	(void)key;
-	(void)value;
-}
-
-void	HttpRequest::set_service_worker_navigation_preload(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_set_cookie(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueMap(value);
-}
-
-void	HttpRequest::set_sourcemap(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_strict_transport_security(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueMap(value);
-}
-
-void	HttpRequest::set_te(const std::string &key, const std::string &value)
-{
-	std::stringstream				splited_by_commma(value);
-	std::string						line;
-	std::string						target_key;
-	std::string						target_value;
-
-	while(std::getline(splited_by_commma, line, ','))
-	{
-		if (line.find(';') != std::string::npos)
-		{
-			if (line[0] == ' ')
-				line = line.substr(1);
-			target_key = HandlingString::obtain_word_before_delimiter(line, ';');
-			target_value = HandlingString::obtain_weight(HandlingString::obtain_word_after_delimiter(line, ';'));
-			if (!(target_key == "compress" || target_key == "deflate" || target_key == "gzip" || target_key == "trailers"))
-				return;
-			if (HandlingString::is_double(target_value) == false)
-				return;
-		}
-		else
-		{
-			if (line[0] == ' ')
-				line = line.substr(1);
-			if (!(line == "compress" || line == "deflate" || line == "gzip" || line == "trailers"))
-				return;
-		}
-	}
-	this->_request_keyvalue_map[key] = ready_ValueWeightArraySet(value);
-}
-
-void	HttpRequest::set_timing_allow_origin(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_trailer(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_transfer_encoding(const std::string &key, const std::string &value)
-{
-	std::stringstream	ss(value);
-	std::string			line;
-	std::string			skipping_firstemptyword;
-
-	while(std::getline(ss, line, ','))
-	{
-		if (line[0] == ' ')
-			skipping_firstemptyword = line.substr(1);
-		else
-			skipping_firstemptyword = line;
-		if (skipping_firstemptyword != "gzip" && skipping_firstemptyword != "compress" && skipping_firstemptyword \
-		!= "deflate" && skipping_firstemptyword != "gzip" && skipping_firstemptyword != "chunked")
-			return;
-	}
-	this->_request_keyvalue_map[key] = ready_ValueArraySet(value);
-}
-
-void	HttpRequest::set_upgrade(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueArraySet(value);
-}
-
-void	HttpRequest::set_upgrade_insecure_requests(const std::string &key, const std::string &value)
-{
-	if (HandlingString::is_positive_and_under_intmax(value) == false)
-		return;
-	if (HandlingString::is_positive_and_under_intmax(value) == false)
-		return;
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_user_agent(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_vary(const std::string &key, const std::string &value)
-{
-	// headerのみしか許可しないのでは
-	this->_request_keyvalue_map[key] = ready_ValueArraySet(value);
-}
-
-void	HttpRequest::set_via(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueSet(value);
-}
-
-void	HttpRequest::set_www_authenticate(const std::string &key, const std::string &value)
-{
-	this->_request_keyvalue_map[key] = ready_ValueArraySet(value);
-}
-
-void	HttpRequest::set_x_xss_protection(const std::string &key, const std::string &value)
-{
-	(void)key;
-	(void)value;
 }
 
 void HttpRequest::ready_functionmap()
